Migrating List to appropriate component
